requirements.txt
django==2.2
djangorestframework==3.11.0
rest_framework_jwt=1.11.0